+++ packages/terra-alert/CHANGELOG.md
@@ -4,6 +4,7 @@
 
 * Changed
   * Updated jest snapshots.
+  * Updated jest snapshots.
 
 ## 4.60.6 - (May 3, 2022)