+++ packages/terra-status-view/CHANGELOG.md
@@ -4,6 +4,7 @@
 
 * Changed
   * Updated jest snapshots.
+  * Updated jest snapshots.
   
 ## 4.48.5 - (May 3, 2022)